# php/public/openapi.yaml
openapi: 3.0.3
info:
  title: API سامانه کنترل خطوط تاکسیرانی مشهد
  version: 0.5.0
  description: مستندات وب‌سرویس سامانه (احراز هویت، عملیات میدانی، مدیریت و گزارش‌گیری)
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
security:
  - bearerAuth: []
paths:
...
  /official-visits:
    post:
      tags: [officials]
      summary: ثبت حضور یک مسئول در خط
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [official_id]
              properties:
                official_id: { type: integer }
                line_id: { type: integer }
                note: { type: string }
                lat: { type: number }
                lng: { type: number }
      responses: { '201': { description: ثبت شد } }
...
  /admin/users/{id}:
    put: { tags: [admin], summary: ویرایش کاربر, parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: ok } } }
  /admin/users/{id}/lines:
    get: { tags: [admin], summary: خطوط مجاز کاربر, parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: لیست } } }
    put: { tags: [admin], summary: تخصیص خطوط مجاز, parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: ok } } }
  /admin/users/{id}/org:
    put: { tags: [admin], summary: تعیین سرپرست/منطقه (چارت), parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: ok } } }
...
  /admin/report:
    get:
      tags: [admin]
      summary: گزارش‌گیری تجمیعی (خروجی ستون/سطر)
      parameters:
        - { name: type, in: query, schema: { type: string, enum: [attendance, notices, checklists, bills] } }
        - { name: from, in: query, schema: { type: string } }
        - { name: to, in: query, schema: { type: string } }
      responses: { '200': { description: cols + rows } }
...
  /admin/geofences/{id}:
    put: { tags: [admin], summary: ویرایش محدوده, parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: ok } } }
    delete: { tags: [admin], summary: حذف محدوده, parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: ok } } }
...
  /admin/messages/{id}/receipts:
    get: { tags: [admin], summary: رسید خواندن (چه کسانی خواندند/نخواندند), parameters: [{ name: id, in: path, required: true, schema: { type: integer } }], responses: { '200': { description: لیست } } }
...